move tests to their own file
